requirements.txt
flask
firebase-admin
python-dotenv
mysql-connector-python